LayoutTests/fast/selectors/not-nested-expected.txt

Check the validity of :not() inside :not() matching.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing ":not(testcaseA)"
PASS document.querySelectorAll('#test-root :not(testcaseA)').length is 5
PASS document.querySelectorAll('#test-root :not(testcaseA)')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :not(testcaseA)')[1].id is "testcase4"
PASS document.querySelectorAll('#test-root :not(testcaseA)')[2].id is "testcase6"
PASS document.querySelectorAll('#test-root :not(testcaseA)')[3].id is "testcase8"
PASS document.querySelectorAll('#test-root :not(testcaseA)')[4].id is "testcase10"
PASS styledElements is ["testcase2","testcase4","testcase6","testcase8","testcase10"]

Testing ":not(:not(testcaseA))"
PASS document.querySelectorAll('#test-root :not(:not(testcaseA))').length is 5
PASS document.querySelectorAll('#test-root :not(:not(testcaseA))')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root :not(:not(testcaseA))')[1].id is "testcase3"
PASS document.querySelectorAll('#test-root :not(:not(testcaseA))')[2].id is "testcase5"
PASS document.querySelectorAll('#test-root :not(:not(testcaseA))')[3].id is "testcase7"
PASS document.querySelectorAll('#test-root :not(:not(testcaseA))')[4].id is "testcase9"
PASS styledElements is ["testcase1","testcase3","testcase5","testcase7","testcase9"]

Testing ":not(:not(:not(testcaseA)))"
PASS document.querySelectorAll('#test-root :not(:not(:not(testcaseA)))').length is 5
PASS document.querySelectorAll('#test-root :not(:not(:not(testcaseA)))')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :not(:not(:not(testcaseA)))')[1].id is "testcase4"
PASS document.querySelectorAll('#test-root :not(:not(:not(testcaseA)))')[2].id is "testcase6"
PASS document.querySelectorAll('#test-root :not(:not(:not(testcaseA)))')[3].id is "testcase8"
PASS document.querySelectorAll('#test-root :not(:not(:not(testcaseA)))')[4].id is "testcase10"
PASS styledElements is ["testcase2","testcase4","testcase6","testcase8","testcase10"]

Testing ":not(:not(:not(:not(testcaseA))))"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(testcaseA))))').length is 5
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(testcaseA))))')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(testcaseA))))')[1].id is "testcase3"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(testcaseA))))')[2].id is "testcase5"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(testcaseA))))')[3].id is "testcase7"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(testcaseA))))')[4].id is "testcase9"
PASS styledElements is ["testcase1","testcase3","testcase5","testcase7","testcase9"]

Testing ":not(:not(:not(:not(:not(testcaseA)))))"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(testcaseA)))))').length is 5
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(testcaseA)))))')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(testcaseA)))))')[1].id is "testcase4"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(testcaseA)))))')[2].id is "testcase6"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(testcaseA)))))')[3].id is "testcase8"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(testcaseA)))))')[4].id is "testcase10"
PASS styledElements is ["testcase2","testcase4","testcase6","testcase8","testcase10"]

Testing ":not(:not(:not(:not(:not(:not(testcaseA))))))"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(testcaseA))))))').length is 5
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(testcaseA))))))')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(testcaseA))))))')[1].id is "testcase3"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(testcaseA))))))')[2].id is "testcase5"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(testcaseA))))))')[3].id is "testcase7"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(testcaseA))))))')[4].id is "testcase9"
PASS styledElements is ["testcase1","testcase3","testcase5","testcase7","testcase9"]

Testing ":not(:not(:not(:not(:not(:not(:not(testcaseA)))))))"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(testcaseA)))))))').length is 5
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(testcaseA)))))))')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(testcaseA)))))))')[1].id is "testcase4"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(testcaseA)))))))')[2].id is "testcase6"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(testcaseA)))))))')[3].id is "testcase8"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(testcaseA)))))))')[4].id is "testcase10"
PASS styledElements is ["testcase2","testcase4","testcase6","testcase8","testcase10"]

Testing ":not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))').length is 5
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))')[1].id is "testcase3"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))')[2].id is "testcase5"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))')[3].id is "testcase7"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))')[4].id is "testcase9"
PASS styledElements is ["testcase1","testcase3","testcase5","testcase7","testcase9"]

Testing ":not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))').length is 5
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))')[1].id is "testcase4"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))')[2].id is "testcase6"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))')[3].id is "testcase8"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))')[4].id is "testcase10"
PASS styledElements is ["testcase2","testcase4","testcase6","testcase8","testcase10"]

Testing ":not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))').length is 5
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))')[1].id is "testcase3"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))')[2].id is "testcase5"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))')[3].id is "testcase7"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))')[4].id is "testcase9"
PASS styledElements is ["testcase1","testcase3","testcase5","testcase7","testcase9"]

Testing ":not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))').length is 5
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))')[1].id is "testcase4"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))')[2].id is "testcase6"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))')[3].id is "testcase8"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))')[4].id is "testcase10"
PASS styledElements is ["testcase2","testcase4","testcase6","testcase8","testcase10"]

Testing ":not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))))"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))))').length is 5
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))))')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))))')[1].id is "testcase3"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))))')[2].id is "testcase5"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))))')[3].id is "testcase7"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))))')[4].id is "testcase9"
PASS styledElements is ["testcase1","testcase3","testcase5","testcase7","testcase9"]

Testing ":not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))))"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))))').length is 5
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))))')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))))')[1].id is "testcase4"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))))')[2].id is "testcase6"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))))')[3].id is "testcase8"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))))')[4].id is "testcase10"
PASS styledElements is ["testcase2","testcase4","testcase6","testcase8","testcase10"]

Testing ":not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))))))"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))))))').length is 5
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))))))')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))))))')[1].id is "testcase3"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))))))')[2].id is "testcase5"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))))))')[3].id is "testcase7"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))))))')[4].id is "testcase9"
PASS styledElements is ["testcase1","testcase3","testcase5","testcase7","testcase9"]

Testing ":not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))))))"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))))))').length is 5
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))))))')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))))))')[1].id is "testcase4"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))))))')[2].id is "testcase6"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))))))')[3].id is "testcase8"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))))))')[4].id is "testcase10"
PASS styledElements is ["testcase2","testcase4","testcase6","testcase8","testcase10"]

Testing ":not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))))))))"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))))))))').length is 5
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))))))))')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))))))))')[1].id is "testcase3"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))))))))')[2].id is "testcase5"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))))))))')[3].id is "testcase7"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA))))))))))))))))')[4].id is "testcase9"
PASS styledElements is ["testcase1","testcase3","testcase5","testcase7","testcase9"]

Testing ":not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))))))))"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))))))))').length is 5
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))))))))')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))))))))')[1].id is "testcase4"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))))))))')[2].id is "testcase6"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))))))))')[3].id is "testcase8"
PASS document.querySelectorAll('#test-root :not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(:not(testcaseA)))))))))))))))))')[4].id is "testcase10"
PASS styledElements is ["testcase2","testcase4","testcase6","testcase8","testcase10"]

Testing ":not(testcaseA, :not(testcaseA))"
PASS document.querySelectorAll('#test-root :not(testcaseA, :not(testcaseA))').length is 0
PASS styledElements is []

Testing ".baz:not(testcaseA)"
PASS document.querySelectorAll('#test-root .baz:not(testcaseA)').length is 3
PASS document.querySelectorAll('#test-root .baz:not(testcaseA)')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root .baz:not(testcaseA)')[1].id is "testcase6"
PASS document.querySelectorAll('#test-root .baz:not(testcaseA)')[2].id is "testcase8"
PASS styledElements is ["testcase2","testcase6","testcase8"]

Testing ":not(testcaseA:not(.baz))"
PASS document.querySelectorAll('#test-root :not(testcaseA:not(.baz))').length is 9
PASS document.querySelectorAll('#test-root :not(testcaseA:not(.baz))')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root :not(testcaseA:not(.baz))')[1].id is "testcase2"
PASS document.querySelectorAll('#test-root :not(testcaseA:not(.baz))')[2].id is "testcase3"
PASS document.querySelectorAll('#test-root :not(testcaseA:not(.baz))')[3].id is "testcase4"
PASS document.querySelectorAll('#test-root :not(testcaseA:not(.baz))')[4].id is "testcase5"
PASS document.querySelectorAll('#test-root :not(testcaseA:not(.baz))')[5].id is "testcase6"
PASS document.querySelectorAll('#test-root :not(testcaseA:not(.baz))')[6].id is "testcase7"
PASS document.querySelectorAll('#test-root :not(testcaseA:not(.baz))')[7].id is "testcase8"
PASS document.querySelectorAll('#test-root :not(testcaseA:not(.baz))')[8].id is "testcase10"
PASS styledElements is ["testcase1","testcase2","testcase3","testcase4","testcase5","testcase6","testcase7","testcase8","testcase10"]

PASS successfullyParsed is true

TEST COMPLETE
